moved html logic to GenerateFrontend and renamed PrepareIndexHtml to GenerateIndexJS
